examples/md_to_docx_agent/pyproject.toml
[project]
name = "md-to-docx-agent"
version = "0.1.0"
description = "Markdown to DOCX conversion agent example"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-openai>=0.0.8",
    "langchain-core>=0.3.31",
    "middleman-ai",
    "python-dotenv>=1.0.1",
]

[tool.uv]
[tool.uv.sources]
middleman-ai = { path = "../../", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/md_to_docx_agent",
]
